requirements.txt
opencv-python-headless
ultralytics
google-generativeai
numpy
streamlit_webrtc
selenium
aiohttp